Memory: New memory model
Address data type, frontend and backend interface separation, memcpy-like API, refactor

+++ qtmips_gui/lcddisplayview.h
@@ -20,7 +20,7 @@ public:
     uint fb_height();
 
 public slots:
-    void pixel_update(uint x, uint y, uint r, uint g, uint b);
+    void pixel_update(size_t x, size_t y, uint r, uint g, uint b);
 
 protected:
     void paintEvent(QPaintEvent *event) override;
